moving some schemaorg utility classes from the corelib-schemaorg module to the corelib-record module
