+++ corelib-record/src/main/java/eu/europeana/corelib/record/schemaorg/utils/SchemaOrgTypeFactory.java
@@ -1,4 +1,4 @@
-package eu.europeana.corelib.edm.utils;
+package eu.europeana.corelib.record.schemaorg.utils;
 
 import eu.europeana.corelib.definitions.edm.entity.Agent;
 import eu.europeana.corelib.definitions.edm.entity.ContextualClass;
