+++ corelib-record/src/main/java/eu/europeana/corelib/record/schemaorg/utils/SchemaOrgUtils.java
@@ -1,4 +1,4 @@
-package eu.europeana.corelib.edm.utils;
+package eu.europeana.corelib.record.schemaorg.utils;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -17,6 +17,7 @@ import java.util.Map;
 import java.util.Set;
 
 import eu.europeana.corelib.edm.model.schemaorg.*;
+import eu.europeana.corelib.edm.utils.JsonLdSerializer;
 import eu.europeana.corelib.solr.entity.*;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
