+++ corelib-record/src/test/java/eu/europeana/corelib/record/schemaorg/utils/MockEdmOrganization.java
@@ -1,4 +1,4 @@
-package eu.europeana.corelib.edm.utils;
+package eu.europeana.corelib.record.schemaorg.utils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
