+++ corelib-record/src/test/resources/schemaorg/edmorganization.json
