+++ corelib-record/src/test/resources/schemaorg/fullbean.json
