+++ corelib-schemaorg/pom.xml
@@ -17,33 +17,11 @@
     <url>http://www.europeana.eu</url>
 
     <dependencies>
-        <dependency>
-            <groupId>eu.europeana.corelib</groupId>
-            <artifactId>corelib-definitions</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>eu.europeana.corelib</groupId>
-            <artifactId>corelib-storage</artifactId>
-        </dependency>
-
         <!-- Schema.org -->
         <dependency>
             <groupId>com.io-informatics.oss</groupId>
             <artifactId>jackson-jsonld</artifactId>
             <version>0.1.1</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-test</artifactId>
-            <version>${springCore.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-            <version>${mockito.version}</version>
-            <scope>test</scope>
         </dependency>
     </dependencies>
